% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_sur.R
\name{ci_sur}
\alias{ci_sur}
\title{confidence interval for difference in binomial proportions
incorporating clinical experts survey data for clinical non-inferiority
margin}
\usage{
ci_sur(dt_sur, dt_ct, type = c("all", "obs", "mi", "sing"),
  subj_miss = FALSE, num_ms = 10)
}
\arguments{
\item{dt_sur}{tibble, summarised survey data}

\item{dt_ct}{tibble, clinical trial data}

\item{type}{string, Default: c("all", "obs", "mi", "sing"), type of the
survey data used}

\item{subj_miss}{logic, Default: FALSE, detemines whether subject level data
are incomplete}

\item{num_ms}{numeric, Default: 10, number of imputations for subject level
data}
}
\value{
tibble, inlcudes lower/upper confidence interval bounds, and
non-inferiority decision
}
\description{
calculates confidence interval for difference in binomial
proportions using Wald method, while incorporrating clinical experts survey
data in regards to non-inferiority margin
}
\seealso{
\code{\link[dplyr]{mutate}}
 \code{\link[purrr]{map2}}
 \code{\link[bin2mi]{p2_mle}}
 \code{\link[tidyr]{unnest}}
 \code{\link[stats]{Normal}}
}
