% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutval.R
\name{cutval}
\alias{cutval}
\title{calculates a cut-off value to be used in the confidence interval}
\usage{
cutval(alpha = 0.025, df_m2, lambda)
}
\arguments{
\item{alpha}{numeric, Default: 0.025, a one-sided level of alpha to be used}

\item{df_m2}{numeric, degrees of freedom from margin imputation}

\item{lambda}{numeric, ratio between degrees of freedom between margin
imputation and subject level imputation, if subject level data was not
imputed, then set lambda as zero}
}
\value{
numeric, cut-off value
}
\description{
calcualtes cut-off value to be used in the confidence itnerval
}
\examples{
cutval(0.025, 50, 1)
}
\seealso{
\code{\link[stats]{Normal}}
}
