% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_miss_pert.R
\name{dt_miss_pert}
\alias{dt_miss_pert}
\title{Impose missingness}
\usage{
dt_miss_pert(dt, do_ratec, do_ratet, bxmc, bxmt)
}
\arguments{
\item{dt}{tibble, data to impose missingness on}

\item{do_ratec}{numeric, target drop out rate for group 'c'}

\item{do_ratet}{numeric, target drop out rate for group 't'}

\item{bxmc}{numeric, x coefficient for affecting missingness in group 'c'}

\item{bxmt}{numeric, x coefficient for affecting missingness in group 't'}
}
\value{
tibble
}
\description{
imposes missingness per each group seperately
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{n}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{select}}
 \code{\link[stats]{Uniform}}
}
