% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2_mi.R
\name{m2_mi}
\alias{m2_mi}
\title{multiple imputation to set a margin}
\usage{
m2_mi(dt, num_m = 10, maxit = 20, mi_method = "norm",
  use_pckg = "mice", i, n_iter = 100)
}
\arguments{
\item{dt}{tibble, contains clinical experts survey information}

\item{num_m}{numeric, Default: 10, number of imputed datasets}

\item{maxit}{numeris, Default: 20, number of maximum iterations per
imputation}

\item{mi_method}{characgter, Default: "norm", method used for mice mi}

\item{use_pckg}{character, 'mice' or 'norm', package to be used in MI}

\item{i}{numeric, the seed is defined as 666*i (norm package)}

\item{n_iter}{numeric, Default: 100, number of iterations to be used in data
augmentation procedure (norm package)}
}
\value{
tibble, contains summary of the multiply imputed responses
}
\description{
multiply imputes unobserved fraction for preservation of control
treatment effect over placebo, which is used to set up a clincal margin (M2)
for non-inferiority trial
}
\details{
the function reads a partially observed survey data, then multiple
imputes it using mice with normal model
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[purrr]{map2}}
 \code{\link[stats]{cor}}
}
