% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2_vary.R
\name{m2_vary}
\alias{m2_vary}
\title{Variable margin overall point estimate and variance derivation}
\usage{
m2_vary(lambda, m1, pc, pt, pd_var, nc, nt, method = c("wald", "fm"))
}
\arguments{
\item{lambda}{numeric, fraction of control treatment effect preservation}

\item{m1}{numeric, statistical margin}

\item{pc}{numeric, estimated proportion of events in group 'c'}

\item{pt}{numeric, estimated proportion of events in group 't'}

\item{pd_var}{numeric, estimated variance for difference in propotions
between groups 'c' and 't'}

\item{nc}{numeric, number of observations in group 'c'}

\item{nt}{numeric, number of observations in group 't'}

\item{method}{chararcter with the followin two options: "wald", "fm"}
}
\value{
list
}
\description{
Calculates overall point estimates and its variance using a
vector of margins
}
\details{
DETAILS
}
\examples{
lambda <- c(0.60, 0.63, 0.65)
m2_vary(lambda, m1 = 0.23, pc = 0.8, pt = 0.7, pd_var = 0.004, nc = 100, nt = 100, method = 'wald')
}
\seealso{
\code{\link[stats]{cor}}
 \code{\link[purrr]{map}},\code{\link[purrr]{set_names}}
 \code{\link[bin2mi]{p_rmle}}
}
