% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_run.R
\name{norm_run}
\alias{norm_run}
\title{Runs MI via norm package}
\usage{
norm_run(dt_in, num_m = 5, i, n_iter)
}
\arguments{
\item{dt_in}{tibble, to be imputed}

\item{num_m}{numeric, Default: 5, number of imputations}

\item{i}{numeric, the seed is defined as 666*i}

\item{n_iter}{numeric, number of iterations to be used in data
augmentation procedure}
}
\value{
tibble of MI summary: qbar, ubar, b, v and t
}
\description{
Runs MI via norm package
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[purrr]{map2}}
 \code{\link[stats]{cor}}
}
