% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_fm.R
\name{ss_fm}
\alias{ss_fm}
\title{sample size calcukation for difference in binomial proportions}
\usage{
ss_fm(alpha, power, pc, pt, m2, pc_r, pt_r)
}
\arguments{
\item{alpha}{numeric, desired alpha level (one-sided)}

\item{power}{numeric, desirerd power level}

\item{pc}{numeric, assumed proportion of events in group 'c'}

\item{pt}{numeric, assumed proportion of events in group 't'}

\item{m2}{numerric, pre-specified margin}

\item{pc_r}{numeric, restricted under the null proportion of events in group
'c'}

\item{pt_r}{numeric, restricted under the null proportion of events in group
't't}
}
\value{
numeric
}
\description{
sample size calcukation for difference in binomial proportions
using Farrington-Manning method
}
\details{
function returns sample size per group assumimg 1:1 allocation ratio
}
\examples{
ss_fm(0.025, 0.8, 0.8, 0.8, 0.1, 0.78, 0.78 - 0.1)
}
\seealso{
\code{\link[stats]{Normal}}
}
