% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_wald.R
\name{ss_wald}
\alias{ss_wald}
\title{sample size calcukation for difference in binomial proportions}
\usage{
ss_wald(alpha, power, pc, pt, m2)
}
\arguments{
\item{alpha}{numeric, desired alpha level (one-sided)}

\item{power}{numeric, desirerd power level}

\item{pc}{numeric, assumed proportion of events in group 'c'}

\item{pt}{numeric, assumed proportion of events in group 't'}

\item{m2}{numerric, pre-specified margin}
}
\value{
numeric
}
\description{
sample size calcukation for difference in binomial proportions
using Wald method
}
\details{
function returns sample size per group assumimg 1:1 allocation ratio
}
\examples{
ss_wald(0.025, 0.8, 0.8, 0.8, 0.1)
}
