% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subj_mi.R
\name{subj_mi}
\alias{subj_mi}
\title{multiple imputation for subject level data}
\usage{
subj_mi(dt, num_ms = 10, maxit = 20)
}
\arguments{
\item{dt}{tibble, subject level data}

\item{num_ms}{numeric, Default: 10, number of imputed datasets}

\item{maxit}{numeris, Default: 20, number of maximum iterations per
imputation}
}
\value{
tibble, contains summary of the multiply imputed responses
}
\description{
multiply imputes unobserved subject level data using mice
}
\details{
the function reads a partially observed subject level data,
then multiple imputes it using mice with logistic model
}
\seealso{
\code{\link[tibble]{tibble}}
 \code{\link[purrr]{map2}}
 \code{\link[stats]{cor}}
}
